refactor : int price -> long price 로 변경

+++ src/main/java/com/lesson/spring/motel/api/request/MotelCreateRequest.java
@@ -6,9 +6,9 @@ import lombok.Getter;
 public class MotelCreateRequest {
 
     private final String name;
-    private final int price;
+    private final long price;
 
-    public MotelCreateRequest(String name, int price) {
+    public MotelCreateRequest(String name, long price) {
         this.name = name;
         this.price = price;
     }
